HUGEE UPDATEEEE
This is just a refreshment for everything. Things look cleaner but ALL THE BLOGS ARE JUST A DEMO COPIED FROM WEBSITES "DONT TAKE IT SERIOUSLY"
And a contact forum which is just there and doesnt do much, but its there :D Home page is also designed
In the next update i will make another menu tab for a RANDOM CHAMPION GENERATOR for league of legends. Its mostly for personal use but still fun since it my first time messing with js.
well thats it.